Calculate extreme shifts

calc_extreme_shifts is used to calculate the minimum and maximum shifts, whilst preserving the criteria that at least min_num_overlapping_points are being compared from both accessions.

calc_extreme_shifts(
  mean_df,
  min_num_overlapping_points,
  shift_extreme,
  accession_data_to_transform,
  accession_data_ref
)

Arguments

mean_df

Input dataframe containing data to transform and reference data.

min_num_overlapping_points

Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.

shift_extreme

Approximation of maximum and minimum shifts allowed.

accession_data_to_transform

Accession name of data which will be transformed.

accession_data_ref

Accession name of reference data.

Value

Minimum and maximum values of shifts.