Calculate best shifts and stretches for each gene, also calculate AIC/BIC under registration or non-registration

get_best_stretch_and_shift is a function to stretch in all stretches and calculates best shift, by comparing SUM of squares difference. For the best shift in each stretch, compares to separate models to calculate AIC/BIC under registration or no registration.

get_best_stretch_and_shift(
  to_shift_df,
  all_data_df,
  stretches,
  do_rescale,
  min_num_overlapping_points,
  shift_extreme,
  num_shifts,
  accession_data_to_transform,
  accession_data_ref,
  data_to_transform_time_added,
  data_ref_time_added
)

Arguments

to_shift_df
Input data containing mean of each time point.

all_data_df
Input all data (without taking mean).

stretches
Vector data of stretches.

do_rescale
Apply "scale" to compared points for each shift if TRUE, use original mean expression data if FALSE.

min_num_overlapping_points
Bound the extreme allowed shifts, such than at least this many timepoints are being compared for both accessions.

shift_extreme
Approximation of maximum and minimum shifts allowed.

num_shifts
Number of different shifts to be considered.

accession_data_to_transform
Accession name of data which will be transformed.

accession_data_ref
Accession name of reference data.

data_to_transform_time_added
Time points to be added in data to transform.

data_ref_time_added
Time points to be added in reference data.

Value

List of data frames (a) all_shifts : all the combos of stretching and shifting tried for each gene, (b) best_shifts : the best stretch and shift combo found for each gene, as well as info for scaling, and (c) model_comparison.dt : AIC / BIC scores for best registerd model found, compared to separate model for each genes expression in the 2 accessions.