Get expression of interest

Get expression of interest

Usage

get_expression_of_interest(
  data_ref,
  data_to_transform,
  id_table,
  lookup_col_ref_and_id_table = "CDS.model",
  lookup_col_ref_and_to_transform = "locus_name",
  colnames_wanted = NULL,
  tissue_wanted = NULL,
  gene_of_interest_acc,
  sum_exp_data_ref = FALSE,
  accession_data_to_transform = "Col0"
)

Arguments

data_ref

File name in working directory, path to file of reference data.

data_to_transform

File name in working directory, path to file of data to transform.

id_table

File name in working directory, path to file of ID table connecting both reference data and data.

lookup_col_ref_and_id_table

Column names shared by both reference data and ID table.

lookup_col_ref_and_to_transform

Column names shared by both reference data and data to transform.

colnames_wanted

List of column names to keep from both reference data and data to transform.

tissue_wanted

Name of tissue from which data will be compared.

gene_of_interest_acc

Gene accession list from data to transform.

sum_exp_data_ref

If TRUE then sum all gene data. Default is FALSE.

accession_data_to_transform

Accession name of data which will be transformed.

Value

A data frame contains both reference data and data to transform for selected gene of interest.