Simplify reference data and data to transform into one dataframe

Simplify reference data and data to transform into one dataframe

get_mean_and_all_exp_data(
  filepath_data_ref,
  filepath_data_to_transform,
  filepath_id_table,
  fix_id_table_shared_colname,
  fix_and_to_transform_data_shared_colname,
  colnames_id_table,
  colnames_wanted,
  tissue_wanted,
  curr_GoIs,
  sum_exp_data_ref = FALSE,
  accession_data_to_transform = "Col0",
  ids_data_ref_colnames = c("CDS.model", "locus_name"),
  max_expression_value_wanted = 5,
  exp_threshold = 0.5
)

Arguments

filepath_data_ref
File name in working directory, path to file of reference data.

filepath_data_to_transform
File name in working directory, path to file of data to transform.

filepath_id_table
File name in working directory, path to file of ID table connecting both reference data and data.

fix_id_table_shared_colname
Column names shared by both reference data and ID table.

fix_and_to_transform_data_shared_colname
Column names shared by both reference data and data to transform.

colnames_id_table
ID table column names.

colnames_wanted
List of column names to keep from both reference data and data to transform.

tissue_wanted
Name of tissue from which data will be compared.

curr_GoIs
Gene of interest list.

sum_exp_data_ref
If TRUE then sum all gene data. Default is FALSE.

accession_data_to_transform
Accession name of data which will be transformed.

ids_data_ref_colnames
Column names shared by both reference data and ID table, whose element needs to be unique.

max_expression_value_wanted
Maximum value of expression desired.

exp_threshold
Minimum expression threshold from which expression below the threshold will be removed.

Value

Combined data frame for both reference data and data to transform: (1) only containing mean of the expression, and (2) still contains replicate data.