Registering data

scale_and_register_data is a function to process all data. This includes scaling data before registration, finding and calculate score of optimal shifts and stretches, apply the best shifts and stretches.

scale_and_register_data(
  mean_df,
  all_data_df,
  stretches,
  shift_extreme,
  num_shifts,
  min_num_overlapping_points,
  initial_rescale,
  do_rescale,
  accession_data_to_transform,
  accession_data_ref,
  data_to_transform_time_added,
  data_ref_time_added
)

Arguments

mean_df
Input data frame contains the mean gene expression of each gene in each genotype at each timepoint.

all_data_df
Input data frame contains all replicates of gene expression in each genotype at each timepoint.

stretches
Candidate registration stretch factors to apply to data to transform.

shift_extreme
The absolute maximum value which can be applied as a shift to gene expression timecourse (days).

num_shifts
Number of shifts between minimum and maximum values of shift.

min_num_overlapping_points
Number of minimum overlapping time points. Shifts will be only considered if it leaves at least these many overlapping points after applying the registration function.

initial_rescale
Scaling gene expression prior to registration if TRUE.

do_rescale
Scaling gene expression using only overlapping timepoints points during registration.

accession_data_to_transform
Accession name of data which will be transformed.

accession_data_ref
Accession name of reference data.

data_to_transform_time_added
Time points to be added in data to transform.

data_ref_time_added
Time points to be added in reference data.

Value

List of dataframes: (a) mean_df is unchanged by scale_and_register_data(), (b) mean_df_sc is identical to mean_df, with additional column sc.expression_value, (c) imputed_mean_df is registered expression data, (d) all_shifts is a table of candidate registrations applied, and score for each, and (e) model_comparison_dt is a table comparing the optimal registration function for each gene (based on all_shifts scores) to model with no registration applied.